+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Added
 - Method ``SIPDigitalObject.generate_technical_csv_metadata()`` to generate CSV metadata
 - Generate stream metadata in ``SIPDigitalObject.generate_technical_metadata()``
 - Method ``SIP.from_directory`` to generate a SIP object according to the contents of a directory
+- Generate checksum calculation events in ``SIPDigitalObject.generate_technical_metadata()``
 
 Changed
 ^^^^^^^