+++ CHANGELOG.rst
@@ -5,8 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-Unreleased
-----------
+0.1.0 - 2024-03-27
+------------------
 Added
 ^^^^^
 - Method ``SIPDigitalObject.generate_technical_csv_metadata()`` to generate CSV metadata